Postcode_Application/requirements.txt
asgiref==3.2.10
certifi==2020.6.20
chardet==3.0.4
defusedxml==0.6.0
dict2xml==1.7.0
Django==3.1
djangorestframework==3.11.1
djangorestframework-xml==2.0.0
idna==2.10
pytz==2020.1
requests==2.24.0
sqlparse==0.3.1
urllib3==1.25.10
